Fallback to ctime when mtime is not available
as can be the case with the nix store, when checking if command file was
updated. I know ctime is the inode change time of the file, not its
"creation" time, but the birth time is not standard nor available via
zstat, so ctime is the best alternative we have. Also, the nix store
should create a new subdirectory with new files for a new package
version.

Fixes #1